spotter/requirements.txt
requests
pendulum
pyyaml
bs4
pika
tweepy